\name{NaArray-aperm}

\alias{NaArray-aperm}
\alias{NaArray_aperm}
\alias{NaArray-transposition}
\alias{NaArray_transposition}

\alias{t.NaMatrix}
\alias{t,NaMatrix-method}
\alias{aperm.NaArray}
\alias{aperm,NaArray-method}

\title{NaArray transposition}

\description{
  Transpose an \link{NaArray} object by permuting its dimensions.

  WORK-IN-PROGRESS
}

\value{
  COMING SOON...
}

\seealso{
  \itemize{
    \item \code{\link[base]{aperm}()} in base R.

    \item \link{NaArray} objects.

    \item Ordinary \link[base]{array} objects in base R.
  }
}

\examples{
# COMING SOON...
}
\keyword{array}
\keyword{methods}
