\name{SparseArray-aperm}

\alias{SparseArray-aperm}
\alias{SparseArray_aperm}
\alias{SparseArray-transposition}
\alias{SparseArray_transposition}

\alias{t.SVT_SparseMatrix}
\alias{t,SVT_SparseMatrix-method}
\alias{aperm.COO_SparseArray}
\alias{aperm,COO_SparseArray-method}
\alias{aperm.SVT_SparseArray}
\alias{aperm,SVT_SparseArray-method}

\title{SparseArray transposition}

\description{
  Transpose a \link{SparseArray} object by permuting its dimensions.

  WORK-IN-PROGRESS
}

\value{
  COMING SOON...
}

\seealso{
  \itemize{
    \item \code{\link[base]{aperm}()} in base R.

    \item \link{SparseArray} objects.

    \item Ordinary \link[base]{array} objects in base R.
  }
}

\examples{
# COMING SOON...
}
\keyword{array}
\keyword{methods}
