\name{SparseArray-Complex-methods}

\alias{SparseArray-Complex-methods}
\alias{SparseArray_Complex_methods}
\alias{SparseArray-Complex}
\alias{SparseArray_Complex}

\alias{Complex,SVT_SparseArray-method}

\title{'Complex' methods for SparseArray objects}

\description{
  WORK-IN-PROGRESS
}

\value{
  COMING SOON...
}

\seealso{
  \itemize{
    \item \code{\link[methods]{S4groupGeneric}} in the \pkg{methods} package.

    \item \link{SparseArray} objects.

    \item Ordinary \link[base]{array} objects in base R.
  }
}

\examples{
# COMING SOON...
}
\keyword{array}
\keyword{methods}
\keyword{complex}
