\name{SparseArray-dim-tuning}

\alias{SparseArray-dim-tuning}
\alias{SparseArray_dim_tuning}

\alias{tune_Array_dims,SVT_SparseArray-method}

\title{Internal "dim tuning" utilities for SparseArray objects}

\description{
  Internal "dim tuning" utilities for \link{SparseArray} objects not meant
  to be used directly by the end user.
}

\keyword{internal}
