\name{NaArray-subsetting}

\alias{NaArray-subsetting}
\alias{NaArray_subsetting}

\alias{tune_Array_dims,NaArray-method}

\alias{[,NaArray,ANY,ANY,ANY-method}
\alias{extract_na_array}
\alias{extract_na_array,NaArray-method}
\alias{extract_array,NaArray-method}

\title{Subsetting an NaArray object}

\description{
  EXPERIMENTAL!!!

  Like ordinary arrays in base R, \link{NaArray} objects support
  subsetting via the single bracket operator (\code{[}).
}

\seealso{
  \itemize{
    \item \code{\link[base]{drop}} in base R to drop the \emph{ineffective
          dimensions} of an array or array-like object.

    \item \code{\link[S4Arrays]{Lindex2Mindex}} in the \pkg{S4Arrays}
          package for how to convert an \emph{L-index} to an \emph{M-index}
          and vice-versa.

    \item \link{NaArray} objects.

    \item \code{\link[base]{[}} and ordinary \link[base]{array} objects
          in base R.
  }
}

\examples{
# COMING SOON...
}
\keyword{array}
\keyword{methods}
