\name{SparseArray-Math-methods}

\alias{SparseArray_Math_methods}
\alias{SparseArray_Math}

\alias{Math,SVT_SparseArray-method}
\alias{round,SVT_SparseArray-method}
\alias{signif,SVT_SparseArray-method}

\title{'Math' and 'Math2' methods for SparseArray objects}

\description{
  WORK-IN-PROGRESS
}

\value{
  COMING SOON...
}

\seealso{
  \itemize{
    \item \code{\link[methods]{S4groupGeneric}} in the \pkg{methods} package.

    \item \link{SparseArray} objects.

    \item \link[base]{array} objects in base R.
  }
}

\examples{
## COMING SOON...
}
\keyword{array}
\keyword{methods}
\keyword{arith}
