\name{sparseMatrix-utils}

\alias{sparseMatrix-utils}
\alias{sparseMatrix_utils}

\alias{coerce,matrix,dgRMatrix-method}
\alias{coerce,matrix,lgRMatrix-method}

\alias{coerce,Array,dgCMatrix-method}
\alias{coerce,Array,dgRMatrix-method}
\alias{coerce,Array,lgCMatrix-method}
\alias{coerce,Array,lgRMatrix-method}
\alias{coerce,Array,CsparseMatrix-method}
\alias{coerce,Array,RsparseMatrix-method}
\alias{coerce,Array,sparseMatrix-method}

\title{Internal utilities to handle sparseMatrix derivatives}

\description{
  The \pkg{SparseArray} package defines some utilities to handle
  sparseMatrix derivatives (e.g. dgCMatrix and lgCMatrix objects)
  from the \pkg{Matrix} package. These are for internal use only.
}

\seealso{
  \itemize{
    \item \linkS4class{dgCMatrix} objects implemented in the \pkg{Matrix}
          package.
  }
}

\keyword{internal}
\keyword{array}
\keyword{methods}
